# openapi.yaml
components:
  schemas:
    Error:
      properties:
        code:
          type: integer
        error:
          type: string
        message:
          type: string
      type: object
    HealthCheck:
      properties:
        status:
          enum:
          - ok
          - error
          type: string
        timestamp:
          format: date-time
          type: string
        version:
          type: string
      type: object
info:
  description: Auto-generated OpenAPI spec for testql
  title: testql API
  version: 1.0.0
openapi: 3.0.3
paths:
  /oql/file:
    get:
      operationId: oql_read_file
      parameters:
      - in: query
        name: path
        required: false
        schema:
          type: str
      responses:
        '200': &id001
          content:
            application/json:
              schema:
                type: object
          description: Success
        '401': &id002
          description: Unauthorized
        '404': &id003
          description: Not Found
        '500': &id004
          description: Internal Server Error
      summary: Read a TestQL file content (.testql.toon.yaml / .oql / .tql).
      tags:
      - fastapi
      - oql
  /oql/files:
    get:
      operationId: oql_list_files
      responses:
        '200': *id001
        '401': *id002
        '404': *id003
        '500': *id004
      summary: List all .testql.toon.yaml files in the project (with .oql/.tql fallback).
      tags:
      - fastapi
      - oql
  /oql/log:
    get:
      operationId: oql_read_log
      parameters:
      - in: query
        name: name
        required: false
        schema:
          type: str
      responses:
        '200': *id001
        '401': *id002
        '404': *id003
        '500': *id004
      summary: Read a specific log file.
      tags:
      - fastapi
      - oql
  /oql/logs:
    get:
      operationId: oql_list_logs
      responses:
        '200': *id001
        '401': *id002
        '404': *id003
        '500': *id004
      summary: List available log files.
      tags:
      - fastapi
      - oql
  /oql/run-file:
    post:
      operationId: oql_run_file
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        '201': &id005
          content:
            application/json:
              schema:
                type: object
          description: Created
        '400': &id006
          content:
            application/json:
              schema:
                properties:
                  detail:
                    type: string
                  error:
                    type: string
                type: object
          description: Bad Request
        '401': *id002
        '404': *id003
        '500': *id004
      summary: Run an entire OQL file with validation. Returns structured results
        + saves log.
      tags:
      - fastapi
      - oql
  /oql/run-line:
    post:
      operationId: oql_run_line
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        '201': *id005
        '400': *id006
        '401': *id002
        '404': *id003
        '500': *id004
      summary: Execute a single OQL command line via the encoder bridge.
      tags:
      - fastapi
      - oql
  /oql/tables:
    get:
      operationId: oql_list_tables
      parameters:
      - in: query
        name: path
        required: false
        schema:
          type: str
      responses:
        '200': *id001
        '401': *id002
        '404': *id003
        '500': *id004
      summary: Extract table names from an OQL file.
      tags:
      - fastapi
      - oql
servers:
- description: Local development
  url: http://localhost:8101
- description: Relative
  url: /